+++ requirements.txt
@@ -7,3 +7,5 @@ PyAudio==0.2.14
 python-dotenv==1.0.1
 requests==2.31.0
 urllib3==2.2.1
+wave==0.0.2
+pytz==2024.1
